wikidata_sparql_test/requirements.txt
SPARQLWrapper
python-dateutil